Implement the reqired changes
