--- modules/local-storage.js
@@ -1,31 +1,0 @@
-// store data in local storage
-export default class store {
-  static getBooks() {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addBook(book) {
-    const books = store.getBooks();
-
-    books.push(book);
-
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static removeBook(authore) {
-    const books = store.getBooks();
-    books.forEach((book, index) => {
-      if (book.authore === authore) {
-        books.splice(index, 1);
-      }
-    });
-
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-}
